refactor(core): analyze new user counts with 'users' table 'created_at' column
